moved filter logic to store selector/helper
