+++ package.json
@@ -18,7 +18,8 @@
     "react-scripts": "^4.0.3",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.1"
+    "redux-thunk": "^2.4.1",
+    "use-debounce": "^8.0.4"
   },
   "devDependencies": {
     "@cypress/react": "^5.12.4",
